requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-decouple==3.8
python-dotenv>=1.0.0
whitenoise==6.12.0
Pillow==12.1.1
asgiref==3.11.1
sqlparse==0.5.5
pandas>=1.5.0
scikit-learn>=1.0.0
google-genai==1.69.0
djangorestframework-simplejwt==5.5.0
gunicorn==25.3.0
